+++ src/main/java/com/itchain/samplemsa/samplemsa/order/domain/DeliverStatus.java
@@ -1,0 +1,5 @@
+package com.itchain.samplemsa.samplemsa.order.domain;
+
+public enum DeliverStatus {
+    PREPARE, DELIVERING, DONE
+}
