+++ src/main/java/com/itchain/samplemsa/samplemsa/order/domain/DeliveryService.java
@@ -1,0 +1,5 @@
+package com.itchain.samplemsa.samplemsa.order.domain;
+
+public interface DeliveryService {
+    boolean IsAvailableToCancel(String id);
+}
