+++ src/main/java/com/itchain/samplemsa/samplemsa/order/domain/OrderStatus.java
@@ -1,0 +1,5 @@
+package com.itchain.samplemsa.samplemsa.order.domain;
+
+public enum OrderStatus {
+    SIGNED, CANCELED
+}
