+++ src/main/java/com/itchain/samplemsa/samplemsa/order/domain/ProductService.java
@@ -1,0 +1,5 @@
+package com.itchain.samplemsa.samplemsa.order.domain;
+
+public interface ProductService {
+    public int getStockByProductId(String productId);
+}
