+++ src/main/java/com/itchain/samplemsa/samplemsa/order/domain/event/OrderDeletedEvent.java
@@ -1,0 +1,16 @@
+package com.itchain.samplemsa.samplemsa.order.domain.event;
+
+import com.itchain.samplemsa.samplemsa.common.Event;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class OrderDeletedEvent implements Event {
+    private String id;
+
+    @Override
+    public String getID() {
+        return id;
+    }
+}
