+++ src/main/java/com/itchain/samplemsa/samplemsa/order/port/adapter/service/DeliveryAdapter.java
@@ -1,0 +1,7 @@
+package com.itchain.samplemsa.samplemsa.order.port.adapter.service;
+
+import com.itchain.samplemsa.samplemsa.order.domain.DeliverStatus;
+
+public interface DeliveryAdapter {
+    public DeliverStatus getDeliveryStatusByTradeId(String id);
+}
