+++ src/main/java/com/itchain/samplemsa/samplemsa/order/port/adapter/service/ProductAdapter.java
@@ -1,0 +1,5 @@
+package com.itchain.samplemsa.samplemsa.order.port.adapter.service;
+
+public interface ProductAdapter {
+    public int getStockByProductId(String productId);
+}
