--- src/main/java/com/itchain/samplemsa/samplemsa/trade/domain/DeliverStatus.java
@@ -1,5 +1,0 @@
-package com.itchain.samplemsa.samplemsa.trade.domain;
-
-public enum DeliverStatus {
-    PREPARE, DELIVERING, DONE
-}
