--- src/main/java/com/itchain/samplemsa/samplemsa/trade/domain/DeliveryService.java
@@ -1,5 +1,0 @@
-package com.itchain.samplemsa.samplemsa.trade.domain;
-
-public interface DeliveryService {
-    boolean IsAvailableToCancel(String id);
-}
