--- src/main/java/com/itchain/samplemsa/samplemsa/trade/domain/ProductService.java
@@ -1,5 +1,0 @@
-package com.itchain.samplemsa.samplemsa.trade.domain;
-
-public interface ProductService {
-    public int getStockByProductId(String productId);
-}
