--- src/main/java/com/itchain/samplemsa/samplemsa/trade/domain/TradeStatus.java
@@ -1,5 +1,0 @@
-package com.itchain.samplemsa.samplemsa.trade.domain;
-
-public enum TradeStatus {
-    SIGNED, CANCELED
-}
