--- src/main/java/com/itchain/samplemsa/samplemsa/trade/domain/ValidateService.java
@@ -1,15 +1,0 @@
-package com.itchain.samplemsa.samplemsa.trade.domain;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class ValidateService {
-
-    public boolean IsAvailableToCreate(int requestedAmount, int stock) {
-        if (requestedAmount <= stock) {
-            return true;
-        } else {
-            return false;
-        }
-    }
-}
