--- src/main/java/com/itchain/samplemsa/samplemsa/trade/domain/dto/DeliveryInfoDTO.java
@@ -1,14 +1,0 @@
-package com.itchain.samplemsa.samplemsa.trade.domain.dto;
-
-import com.itchain.samplemsa.samplemsa.trade.domain.DeliverStatus;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@AllArgsConstructor
-@Data
-public class DeliveryInfoDTO {
-    private String id;
-    private String productId;
-    private DeliverStatus Status;
-
-}
