--- src/main/java/com/itchain/samplemsa/samplemsa/trade/port/adapter/service/DeliveryAdapter.java
@@ -1,7 +1,0 @@
-package com.itchain.samplemsa.samplemsa.trade.port.adapter.service;
-
-import com.itchain.samplemsa.samplemsa.trade.domain.DeliverStatus;
-
-public interface DeliveryAdapter {
-    public DeliverStatus getDeliveryStatusByTradeId(String id);
-}
