--- src/main/java/com/itchain/samplemsa/samplemsa/trade/port/adapter/service/ProductAdapter.java
@@ -1,7 +1,0 @@
-package com.itchain.samplemsa.samplemsa.trade.port.adapter.service;
-
-import com.itchain.samplemsa.samplemsa.trade.domain.dto.ProductInfoDTO;
-
-public interface ProductAdapter {
-    public int getStockByProductId(String productId);
-}
