+++ src/main/resources/application.yml
@@ -7,4 +7,3 @@ env:
     - delivery
     - customer
     - order
-
