+++ src/test/java/com/itchain/samplemsa/samplemsa/Order/ServiceConfiguration.java
@@ -1,0 +1,36 @@
+package com.itchain.samplemsa.samplemsa.Order;
+
+
+import com.itchain.samplemsa.samplemsa.customer.domain.CustomerService;
+import com.itchain.samplemsa.samplemsa.mock.mockCustomerService;
+import com.itchain.samplemsa.samplemsa.mock.mockDeliveryService;
+import com.itchain.samplemsa.samplemsa.mock.mockProductService;
+import com.itchain.samplemsa.samplemsa.order.domain.DeliveryService;
+import com.itchain.samplemsa.samplemsa.order.domain.ProductService;
+import org.mockito.Mockito;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
+import org.springframework.context.annotation.Profile;
+
+@Profile("test")
+@Configuration
+public class ServiceConfiguration {
+
+    @Bean
+    @Primary
+    public ProductService productService() {
+        return Mockito.mock(mockProductService.class);
+    }
+
+    @Bean
+    @Primary
+    public DeliveryService deliveryService() {
+        return Mockito.mock(mockDeliveryService.class);
+    }
+
+    @Bean
+    @Primary
+    public CustomerService customerService() { return Mockito.mock(mockCustomerService.class);}
+
+}
